Ensure generated DOT file doesn't change unless state machine changes
By using slices and sorting them by state and trigger values we ensure
that ToGraph() will always return the same output for the same state
machine. This is important in situations where ToGraph() is used to
generate a file that's committed to a repository to ensure it doesn't
change on every commit.